functions/requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cycler==0.11.0
Deprecated==1.2.13
docx2txt==0.8
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.37.1
gcloud==0.18.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3.1
jws==0.1.3
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.1
oauth2client==4.1.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-firebase==1.2
python-jwt==3.3.3
python-pptx==0.6.21
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
scipy==1.9.1
seaborn==0.11.2
six==1.16.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
XlsxWriter==3.0.3
